Read A Specific Item Or Child Items Under A Specific Item

Read A Specific Item Or Child Items Under A Specific Item

ztr_read_item(zotero_object, item, child = FALSE, ..., timeout = 20)

Arguments

zotero_object
Zotero object.

item
Item key.

child
Specify whether to retrieve child items under a specific item. Otherwise, a specific item will be retrieved.

...
Query inputs. See Read Requests, Searching, and Sorting and Pagination in the Zotero Web API Documents for available query.

timeout
Seconds for the timeout of request (default: 20).

Value

httr::response object

Examples

if (FALSE) {
ztr <- Zotero$new("ztr4r")

# simple example
ztr %>%
  ztr_read_item(item = "foo")

# example with query: child items under a specific item
ztr %>%
  ztr_read_item(
    item = "foo", child = TRUE,
    format = "json", sort = "dateModified", direction = "asc", limit = 25, start = 0
  )
}