Read Tags Assigned To Items (In A Given Collection)

Read

All Tags In The Library, With The Ability To Filter Based On The Items

Tags Assigned To Top-Level Items

Tags Assigned To Items In A Given Collection

Tags Assigned To Top-Level Items In A Given Collection

ztr_read_items_tags(
  zotero_object,
  collection = NULL,
  top = FALSE,
  ...,
  timeout = 20
)

Arguments

zotero_object
Zotero object.

collection
Collection key (default: NULL).

top
Item level (default: FALSE):

FALSE: All tags assigned to the items in the library/specific collection

TRUE : All tags assigned to the top-level items in the library/specific collection

...
Query inputs. See Read Requests and Sorting and Pagination in the Zotero Web API Documents for available query.

timeout
Seconds for the timeout of request (default: 20).

Value

httr::response object

Examples

if (FALSE) {
ztr <- Zotero$new("ztr4r")

# simple example
ztr %>%
  ztr_read_items_tags()

# example with query: all tags in the library
ztr %>%
  ztr_read_items_tags(
    collection = NULL, top = FALSE,
    format = "json", limit = 25
  )

# example with query: tags assigned to top-level items
ztr %>%
  ztr_read_items_tags(
    collection = NULL, top = TRUE,
    format = "json", limit = 25
  )

# example with query: tags assigned to items in a given collection
ztr %>%
  ztr_read_items_tags(
    collection = "foo", top = FALSE,
    format = "json", limit = 25
  )

# example with query: tags assigned to top-level items in a given collection
ztr %>%
  ztr_read_items_tags(
    collection = "foo", top = TRUE,
    format = "json", limit = 25
  )
}